face_recognition/requirements.txt
face-recognition==1.3.0
opencv-python==4.8.1.78
numpy==1.24.3
pillow==10.0.0
click==8.1.7
